Made the projects section of the course_admin component into its own component and wrote tests to check its functionality.
